requirements.txt
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
authlib==1.6.3
bitarray==3.7.1
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.7
cytoolz==1.0.1
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
eth-abi==5.2.0
eth-account==0.10.0
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-tester==0.10.0b4
eth-typing==5.2.1
eth-utils==5.3.1
fastapi==0.116.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
frozenlist==1.7.0
google-adk==1.14.1
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.113.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.37.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.37.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
mcp==1.14.0
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.3
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
parsimonious==0.10.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
py-ecc==8.0.0
py-env==0.0.1
py-evm==0.12.1b1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyunormalize==16.0.0
pywin32==311
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rlp==4.1.0
rpds-py==0.27.1
rsa==4.9.1
semantic-version==2.10.0
sentry-sdk==2.38.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
tenacity==8.5.0
toolz==1.0.0
trie==3.1.0
typer==0.19.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
web3==6.20.1
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
